feat: перевел react-query на сервер на странице test/id
